test : 도메인별 테스트코드 추가
