+++ src/test/java/racingcar/ValidationUtilTest.java
@@ -1,0 +1,31 @@
+package racingcar;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
+import racingcar.util.ValidationUtil;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class ValidationUtilTest {
+    private ValidationUtil ValidationUtil;
+
+    @BeforeEach
+    void createInstance() {
+        ValidationUtil = new ValidationUtil();
+    }
+
+    @DisplayName("입력 값이 숫자가 아닌 값 일때 false 반환")
+    @ParameterizedTest
+    @CsvSource(value = {"1:true" , "d:false" , "1:true"}, delimiter = ':')
+    void inputNumber(String input, boolean result) {
+        //GIVE
+
+        //WHEN
+
+        //THEN
+        assertThat(ValidationUtil.isNumber(input)).isEqualTo(result);
+    }
+
+}
